+++ pkg/sqlvibe/ext_math.go
@@ -1,0 +1,5 @@
+//go:build SVDB_EXT_MATH
+
+package sqlvibe
+
+import _ "github.com/cyw0ng95/sqlvibe/ext/math"